// home/prismicio-types.d.ts
type Simplify<T> = { [KeyType in keyof T]: T[KeyType] };

type BlogPageDocumentDataSlicesSlice =
  | TextOnlySlice
  | HeroSlice
  | CallToActionSlice
  | AlternateGridSlice;

/**
 * Content for Blog Page documents
 */
interface BlogPageDocumentData {
  /**
   * Slice Zone field in *Blog Page*
   *
   * - **Field Type**: Slice Zone
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_page.slices[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#slices
   */
  slices: prismic.SliceZone<BlogPageDocumentDataSlicesSlice> /**
   * Meta Description field in *Blog Page*
   *
   * - **Field Type**: Text
   * - **Placeholder**: A brief summary of the page
   * - **API ID Path**: blog_page.meta_description
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */;
  meta_description: prismic.KeyTextField;

  /**
   * Meta Image field in *Blog Page*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_page.meta_image
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/field#image
   */
  meta_image: prismic.ImageField<never>;

  /**
   * Meta Title field in *Blog Page*
   *
   * - **Field Type**: Text
   * - **Placeholder**: A title of the page used for social media and search engines
   * - **API ID Path**: blog_page.meta_title
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  meta_title: prismic.KeyTextField;

  /**
   * Card Type field in *Blog Page*
   *
   * - **Field Type**: Select
   * - **Placeholder**: *None*
   * - **Default Value**: Not on Homepage
   * - **API ID Path**: blog_page.card_type
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/field#select
   */
  card_type: prismic.SelectField<
    "Not on Homepage" | "Large Card" | "Small Card",
    "filled"
  >;
}

/**
 * Blog Page document from Prismic
 *
 * - **API ID**: `blog_page`
 * - **Repeatable**: `true`
 * - **Documentation**: https://prismic.io/docs/custom-types
 *
 * @typeParam Lang - Language API ID of the document.
 */
export type BlogPageDocument<Lang extends string = string> =
  prismic.PrismicDocumentWithUID<
    Simplify<BlogPageDocumentData>,
    "blog_page",
    Lang
  >;